moved utilities out of the fixtures module
